test: fix 'long long' format warnings

+++ test/test-get-memory.c
@@ -26,9 +26,9 @@ TEST_IMPL(get_memory) {
   uint64_t free_mem = uv_get_free_memory();
   uint64_t total_mem = uv_get_total_memory();
 
-  printf("free_mem=%llu, total_mem=%llu\n",
-         (unsigned long long) free_mem,
-         (unsigned long long) total_mem);
+  printf("free_mem=%lu, total_mem=%lu\n",
+         (unsigned long) free_mem,
+         (unsigned long) total_mem);
 
   ASSERT(free_mem > 0);
   ASSERT(total_mem > 0);
